requirement.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
python-dotenv
langchain-google-genai
streamlit


######## RAG  Local
langchain-chroma
chromadb


######## RAG Cloud
pypdf
langchain-postgres
psycopg[binary]
ddg

langserve
fastapi
uvicorn
sse_starlette 